add pfdf image assets
