+++ lib/app/widgets/saco_list/saco_list_detail.dart
@@ -6,7 +6,6 @@ import 'package:mapstudio/app/screens/saco/saco_form/saco_form.dart';
 import 'package:mapstudio/app/screens/saco/saco_map/saco_map.dart';
 import 'package:mapstudio/app/widgets/saco_form/saco_accept_request_modal.dart';
 import 'package:mapstudio/app/widgets/saco_list/saco_list_status.dart';
-import 'package:sizer/sizer.dart';
 
 import '../../../common/enums/saco_status_enum.dart';
 import '../../../data/models/pfdf_model.dart';
